socket_select was incorrectly in blocking mode
null being passed as a time out parameter was making it to the socket_select call.  the intention was to have the socket_select call be in non blocking mode by defaulting to zero but passing null in made socket_select blocking.  the result was the parent would hang until a child would finish before being able to process more work.